+++ .github/workflows/main.yml
@@ -23,10 +23,10 @@ jobs:
       - uses: actions/checkout@v3
         with:
           ref: ${{ env.BRANCH_NAME }}
-      - name: Run with setup-python 3.11
+      - name: Run with setup-python 3.8
         uses: actions/setup-python@v4
         with:
-          python-version: '3.11'
+          python-version: '3.8'
           update-environment: true
           cache: 'pipenv'
       - name: Install Selenium
@@ -46,7 +46,7 @@ jobs:
       - name: Install pipenv
         run: pip3 install --user pipenv
       - name: Install dependecies
-        run: pipenv --python 3.11 && pipenv install
+        run: pipenv --python 3.8 && pipenv install
       - name: Build
         run: pipenv run build
       - name: Commit and push if changed
