+++ .github/workflows/release.yml
@@ -14,7 +14,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: '3.11'
+          python-version: '3.8'
           update-environment: true
           cache: 'pipenv'
 
@@ -22,7 +22,7 @@ jobs:
         run: pip install pipenv
 
       - name: Install dependencies with pipenv
-        run: pipenv --python 3.11 && pipenv install
+        run: pipenv --python 3.8 && pipenv install
 
       - name: Install PyInstaller
         run: pipenv install pyinstaller
