+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 更新日志（Changelog）
 
+## v1.2.1
+
+### 2024/6/14
+
+- 兼容 Win7 系统，请使用 Python 版本>=3.8（Compatible with Windows 7 system, please use Python version >= 3.8）
+- 修复部分设备运行更新报错（Fixed an error that occurred when some devices ran updates）
+
 ## v1.2.0
 
 ### 2024/6/9
