+++ config.py
@@ -13,7 +13,7 @@ favorite_list = [
     "湖南卫视",
     "翡翠台",
 ]
-open_online_search = False
+open_online_search = True
 favorite_page_num = 5
 default_page_num = 3
 urls_limit = 15
