+++ version.json
@@ -1,3 +1,3 @@
 {
-  "version": "1.2.0"
+  "version": "1.2.1"
 }
